// application-manager/src/app/shared/interfaces.ts
export interface IApplication {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  age: number;
  phoneNumber: string;
  wayOfCommunication: string;
  englishLevel: string;
  availableToStart: any;
  techSkills: string;
  personalPresentation: string;
  studyFromHome: boolean;
}